pyproject.toml
[project]
name = "autoforge-api"
version = "7.0.0"
description = "AutoForge AI Memory & Audit Engine — SaaS-Ready API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "granian>=1.6.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.6",
    "openai>=1.60.0",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    "neo4j>=5.27.0",
    "numpy>=2.2.0",
    "tiktoken>=0.8.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.18",
    "prometheus-client>=0.21.0",
    "structlog>=24.4.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
